+++ src/components/InvoiceRow.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { Button } from "react-bootstrap";
 import { BiSolidPencil, BiTrash } from "react-icons/bi";
 import { BsEyeFill } from "react-icons/bs";
-import InvoiceModal from "../components/InvoiceModal";
+import InvoiceModal from "./InvoiceModal";
 import { useDispatch } from "react-redux";
 import { deleteInvoice } from "../redux/invoicesSlice";
